+++ PurchaseManagament.Domain/Entities/MaterialOffer.cs
@@ -6,7 +6,7 @@ namespace PurchaseManagament.Domain.Entities
     {
         public long OfferId { get; set; }
         public long MaterialId { get; set; }
-        public double OfferedPrice { get; set; }
+        public decimal OfferedPrice { get; set; }
 
         public virtual Material Material { get; set; }
         public virtual Offer Offer { get; set; }
